+++ e2e/grpc-example/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@e2e/grpc",
+  "type": "module",
+  "private": true,
+  "dependencies": {
+    "@graphql-mesh/compose-cli": "workspace:*",
+    "@graphql-mesh/serve-cli": "workspace:*",
+    "@grpc/grpc-js": "^1.12.2",
+    "@grpc/proto-loader": "^0.7.13",
+    "@omnigraph/grpc": "workspace:*",
+    "graphql": "16.9.0"
+  }
+}
