+++ packages/fusion/runtime/src/federation/supergraph.ts
@@ -1,6 +1,7 @@
 import {
   isEnumType,
   Kind,
+  parse,
   print,
   visit,
   type GraphQLSchema,
