+++ packages/loaders/grpc/src/directives.ts
@@ -2,7 +2,6 @@ import {
   DirectiveLocation,
   GraphQLBoolean,
   GraphQLDirective,
-  GraphQLScalarType,
   GraphQLString,
 } from 'graphql';
 import { ObjMapScalar } from '@graphql-mesh/transport-common';
@@ -11,6 +10,9 @@ export const grpcMethodDirective = new GraphQLDirective({
   name: 'grpcMethod',
   locations: [DirectiveLocation.FIELD_DEFINITION],
   args: {
+    subgraph: {
+      type: GraphQLString,
+    },
     rootJsonName: {
       type: GraphQLString,
     },
@@ -30,6 +32,9 @@ export const grpcConnectivityStateDirective = new GraphQLDirective({
   name: 'grpcConnectivityState',
   locations: [DirectiveLocation.FIELD_DEFINITION],
   args: {
+    subgraph: {
+      type: GraphQLString,
+    },
     rootJsonName: {
       type: GraphQLString,
     },
@@ -43,6 +48,9 @@ export const EnumDirective = new GraphQLDirective({
   name: 'enum',
   locations: [DirectiveLocation.ENUM_VALUE],
   args: {
+    subgraph: {
+      type: GraphQLString,
+    },
     value: {
       type: GraphQLString,
     },
@@ -53,6 +61,9 @@ export const grpcRootJsonDirective = new GraphQLDirective({
   name: 'grpcRootJson',
   locations: [DirectiveLocation.OBJECT],
   args: {
+    subgraph: {
+      type: GraphQLString,
+    },
     name: {
       type: GraphQLString,
     },
