+++ packages/transports/http/src/index.ts
@@ -100,4 +100,4 @@ export default {
 
     return httpExecutor;
   },
-} satisfies Transport;
+} satisfies Transport<HTTPTransportOptions>;
